Counterfactuals Class

A Counterfactuals object should be created by CounterfactualMethodRegr or CounterfactualMethodClassif. It contains the found counterfactuals and has several methods for their evaluation and visualization.

References

Gower, J. C. (1971), "A general coefficient of similarity and some of its properties". Biometrics, 27, 623–637.

Kaufman, L. and Rousseeuw, P.J. (1990), Finding Groups in Data: An Introduction to Cluster Analysis. Wiley, New York.

Active bindings

desired

(list(1) | list(2))
A list with the desired properties of the counterfactuals. For regression tasks it has one element desired_outcome (CounterfactualMethodRegr) and for classification tasks two elements desired_class and desired_prob (CounterfactualMethodRegr).

data

(data.table)
The counterfactuals found.

x_interest

(data.table(1))
A single row with the observation of interest.

Methods

Public methods

Counterfactuals$new()

Counterfactuals$evaluate()

Counterfactuals$predict()

Counterfactuals$plot_parallel()

Counterfactuals$plot_freq_of_feature_changes()

Counterfactuals$get_freq_of_feature_changes()

Counterfactuals$plot_surface()

Counterfactuals$print()

Counterfactuals$clone()

Method new()

Creates a new Counterfactuals object. This method should only be called within the $find_counterfactuals methods of CounterfactualMethodRegr and CounterfactualMethodClassif.

Usage

Counterfactuals$new(cfactuals, predictor, x_interest, param_set, desired)

Arguments

cfactuals

(data.table)
The counterfactuals found. Must have the same column names and types as predictor$data$X.

predictor

(Predictor)
The object (created with iml::Predictor$new()) holding the machine learning model and the data.

x_interest

(data.table(1) | data.frame(1))
A single row with the observation of interest.

param_set

(ParamSet)
A ParamSet created by make_paramset based on predictor$data$X.

desired

(list(1) | list(2))
A list with the desired properties of the counterfactuals. For regression tasks it should have one element desired_outcome (CounterfactualMethodRegr) and for classification tasks two elements desired_class and desired_prob (CounterfactualMethodRegr).

Method evaluate()

Evaluates the counterfactuals. It returns the counterfactuals along with evaluation measures in additional columns.

Usage

Counterfactuals$evaluate(
  measures = c("dist_x_interest", "dist_target", "nr_changed", "dist_train"),
  show_diff = FALSE,
  k = 1L,
  weights = NULL
)

Arguments

measures

(character)
The name of one or more evaluation measures. The following measures are available:

dist_x_interest: The distance of a counterfactual to x_interest. It is measured by the extension of the Gower's distance proposed by Kaufman and Rousseeuw (1990) and implemented by gower.dist.

dist_target: The absolute distance of the prediction of a counterfactual to the desired_outcome (regression task) or desired_prob (classification task).

nr_changed: The number of feature changes w.r.t. x_interest.

dist_train: The (possibly weighted) distance to the k nearest training points.

show_diff

(logical(1))
Should the counterfactuals be displayed as their differences from x_interest? (Default is FALSE.) If set to TRUE, numeric features values of x_interest are substracted from the corresponding feature values of the counterfactuals. Non-numeric features are displayed with their actual counterfactuals value. No difference is indicated by NA.

k

(integerish(1))
How many nearest training points should be considered for computing the dist_train measure? Default is 1L.

weights

(numeric(k) | NULL)
How should the k nearest training points be weighted when computing the dist_train measure? If NULL (default) then all k points are weighted equally. If a numeric vector of length k is given, the i-th element determines the weight of the i-th closest point.

Method predict()

Predicts the target values of the counterfactuals.

Usage

Counterfactuals$predict()

Method plot_parallel()

Plots a parallel plot of the (scaled) feature values of the counterfactuals. x_interest is shown in blue.

Usage

Counterfactuals$plot_parallel(
  feature_names = NULL,
  row_ids = NULL,
  digits_min_max = 2L
)

Arguments

feature_names

(character | NULL)
The names of the (numeric) features to plot. If NULL (default) all features are plotted.

row_ids

(integerish | NULL)
The row ids of the counterfactuals to plot. If NULL (default) all counterfactuals are plotted.

digits_min_max

Number of digits for the rounding the minimum and maximum features values. Default is 3L.

Method plot_freq_of_feature_changes()

Plots a bar chart of the frequency of feature changes across all counterfactuals.

Usage

Counterfactuals$plot_freq_of_feature_changes(subset_zero = FALSE)

Arguments

subset_zero

(logical(1))
Should features that have no changes be excluded from the plot? Default is FALSE.

Method get_freq_of_feature_changes()

Returns the frequency of feature changes across all counterfactuals.

Usage

Counterfactuals$get_freq_of_feature_changes(subset_zero = FALSE)

Arguments

subset_zero

(logical(1))
Should features that have no changes be excluded from the plot? Default is FALSE.

Returns

A (named) numeric vector with the frequency of feature changes.

Method plot_surface()

Creates a surface plot for two features. x_interest is represented as bright dot and all counterfactuals that differ from x_interest only in the selected features are represented as black dots.

(numeric, numeric): surface plot

(non-numeric, non-numeric): heatmap

(numeric, non-numeric): line graph

Usage

Counterfactuals$plot_surface(feature_names, grid_size = 250L)

Arguments

feature_names

(character(2))
The names of the features to plot.

grid_size

(integerish(1))
The grid size of the plot. Ignored in case of two non-numeric features. Default is 250L.

Method print()

Prints the Counterfactuals object.

Usage

Counterfactuals$print()

Method clone()

The objects of this class are cloneable with this method.

Usage

Counterfactuals$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.