NICE (Nearest Instance Counterfactual Explanations) for Classification Task

NICE (Brughmans and Martens, 2021) starts the counterfactual search for x_interest by finding its nearest (optionally) correctly classified neighbor x_nn. The distances function that implements the Heterogeneous Euclidean Overlap Method (HEOM) (Wilson and Martinez, 1997) is used to compute the distances.
Once x_nn was found, NICE iteratively replaces feature values of x_interest with the corresponding values of x_nn to optimize some predefined reward function. Available reward functions are sparsity proximity and plausibility.
Our NICE implementation is equivalent to the original version if return_multiple = FALSE, finish_early = TRUE and x_nn_correct_classif = TRUE.

Details

In the first iteration, NICE creates new instances by replacing one feature of x_interest with the corresponding values of x_nn in each instance. Thus, if x_nn differs from x_interest in d features, d instances are created.
Then, the rewards of each instances are calculated with the chosen reward function.
In the second iteration, new instances are created by replacing one feature of the instance with the highest reward in the previous iteration.
If finish_early = TRUE, the algorithm terminates when the predicted probability for the desired_class of the instance with the highest reward, is in desired_prob. If finish_early = FALSE, the algorithm continues until x_nn is recreated.
Once the algorithm terminated, it depends on return_multiple which instances are returned as counterfactuals. If return_multiple = FALSE, then only the instance with the highest reward in the last iteration is returned as counterfactual. If return_multiple = TRUE, then all instances of all iterations whose predicted probability for the desired_class is in desired_prob are returned as counterfactuals.
Note that if finish_early = FALSE and return_multiple = FALSE, then x_nn is returned as single counterfactual.

References

Brughmans, D., & Martens, D. (2021). NICE: An Algorithm for Nearest Instance Counterfactual Explanations. arXiv preprint arXiv:2104.07411.

Wilson, D Randall, and Tony R Martinez. 1997. “Improved Heterogeneous Distance Functions.” Journal of Artificial Intelligence Research 6: 1–34.

Super classes

counterfactuals::CounterfactualMethod -> counterfactuals::CounterfactualMethodClassif -> NICEClassif

Active bindings

x_nn

(logical(1))
The nearest neighbor.

archive

(list())
A list that stores the history of the algorithm. For each algorithm iteration it has one element. A element contains a data.table that stores all feature combinations in this iterations together with their reward values and their predictions.

Methods

Public methods

NICEClassif$new()

NICEClassif$clone()

Inherited methods

counterfactuals::CounterfactualMethod$print()

counterfactuals::CounterfactualMethodClassif$find_counterfactuals()

Method new()

Create a new NICEClassif object.

Usage

NICEClassif$new(
  predictor,
  optimization = "sparsity",
  x_nn_correct_classif = TRUE,
  return_multiple = TRUE,
  finish_early = TRUE
)

Arguments

predictor

(Predictor)
The object (created with iml::Predictor$new()) holding the machine learning model and the data.

optimization

(character(1))
The optimization strategy that determines the reward function. Can be sparsity (default), proximity or plausibility.

x_nn_correct_classif

(logical(1))
Should only correctly classified observations be considered for the nearest neighbor search? Default is TRUE.

return_multiple

(logical(1))
If set to TRUE (default) all created feature combinations with the desired prediction are returned. If set to FALSE only the feature combination with the desired prediction that has the highest reward in the last iteration is returned. For more information, see the details section. (TODO)

finish_early

(logical(1))
Should the algorithm finish after an iteration in which the feature combinations with the highest reward has the desired prediction? Default is TRUE.

Method clone()

The objects of this class are cloneable with this method.

Usage

NICEClassif$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (require("randomForest")) {
  # Train a model
  rf = randomForest(Species ~ ., data = iris)
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find counterfactuals
  nice_classif = NICEClassif$new(predictor)
  cfactuals = nice_classif$find_counterfactuals(
    x_interest = iris[150L, ], desired_class = "versicolor", desired_prob = c(0.5, 1)
  )
  # Print the results
  cfactuals$data
  # Print archive
  nice_classif$archive
}

#> [[1]]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width reward setosa versicolor
#> 1:          5.9         3.2          5.1         1.8  0.136      0      0.168
#> 2:          5.9         3.0          4.8         1.8  0.314      0      0.346
#>    virginica
#> 1:     0.832
#> 2:     0.654
#> 
#> [[2]]
#>    Sepal.Length Sepal.Width Petal.Length Petal.Width reward setosa versicolor
#> 1:          5.9         3.2          4.8         1.8  0.154      0       0.66
#>    virginica
#> 1:      0.34
#>