WhatIf for Classification Tasks

WhatIf selects the n_counterfactual observations from the training data that are most similar to x_interest and have the desired predicted probabilities for the desired class. As training data predictor$data$X is taken.

Details

To compute dissimilarities, the function uses Gower's dissimilarity measure (Gower, 1990), which is implemented by gower_topn. For numerical features, the distances are scaled with the feature value ranges observed in the training data. Only observations with features values between lower and upper are considered as counterfactual candidates.

References

Gower, J. C. (1971), "A general coefficient of similarity and some of its properties". Biometrics, 27, 623–637.

Wexler, J., Pushkarna, M., Bolukbasi, T., Wattenberg, M., Viégas, F., & Wilson, J. (2019). The what-if tool: Interactive probing of machine learning models. IEEE transactions on visualization and computer graphics, 26(1), 56–65.

Super classes

counterfactuals::CounterfactualMethod -> counterfactuals::CounterfactualMethodClassif -> WhatIfClassif

Methods

Public methods

WhatIfClassif$new()

WhatIfClassif$clone()

Inherited methods

counterfactuals::CounterfactualMethod$print()

counterfactuals::CounterfactualMethodClassif$find_counterfactuals()

Method new()

Create a new WhatIfClassif object.

Usage

WhatIfClassif$new(
  predictor,
  n_counterfactuals = 1L,
  lower = NULL,
  upper = NULL
)

Arguments

predictor

(Predictor)
The object (created with iml::Predictor$new()) holding the machine learning model and the data.

n_counterfactuals

(integerish(1))
The number of counterfactuals to be found. Default is 1L.

lower

(numeric() | NULL)
Vector of minimum values for numeric features. If not NULL, it should be named with the corresponding feature names. If NULL (default) lower is taken for each numeric feature as its minimum value in predictor$data$X.

upper

(numeric() | NULL)
Vector of maximum values for numeric features. If not NULL, it should be named with the corresponding feature names. If NULL (default) upper is taken for each numeric feature as its maximum value in predictor$data$X.

Method clone()

The objects of this class are cloneable with this method.

Usage

WhatIfClassif$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (require("randomForest")) {
  # Train a model
  rf = randomForest(Species ~ ., data = iris)
  # Create a predictor object
  predictor = iml::Predictor$new(rf, type = "prob")
  # Find counterfactuals
  wi_classif = WhatIfClassif$new(predictor, n_counterfactuals = 5L)
  cfactuals = wi_classif$find_counterfactuals(
    x_interest = iris[150L, ], desired_class = "versicolor", desired_prob = c(0.5, 1)
  )
  # Print the results
  cfactuals$data
}

#>    Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1:          5.9         3.2          4.8         1.8
#> 2:          6.0         2.7          5.1         1.6
#> 3:          6.7         3.0          5.0         1.7
#> 4:          6.0         2.9          4.5         1.5
#> 5:          5.9         3.0          4.2         1.5