Creates a ParamSet for the features of a data.table.

Creates a ParamSet for the columns of dt. Depending on the class of a column, a different Param is created:

double: ParamDbl

integer: ParamInt

character: ParamFct (with unique values as levels)

factor: ParamFct (with factor levels as levels)

make_param_set(dt, lower = NULL, upper = NULL)

Arguments

dt

(data.table())
The data for the ParamSet.

lower

(numeric() | NULL)
Vector of minimum values for numeric features. If not NULL, it should be named with the corresponding feature names. If NULL (default) lower is taken for each numeric feature as its minimum value in dt.

upper

(numeric() | NULL)
Vector of maximum values for numeric features. If not NULL, it should be named with the corresponding feature names. If NULL (default) upper is taken for each numeric feature as its maximum value in dt.

Value

A ParamSet for the features of dt.